requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
idna==3.4
phonenumbers==8.13.13
libjpeg
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
wheel==0.37.1
six==1.16.0
pymemcache==4.0.0
